DAOS-12555 control: Misc improvements in daos_server storage commands
- Refuse to launch if daos_server is already running
- Create general command types to centralise init
- Add socket ID specifier option to daos_server scm scan
- Remove unused hugepages on daos_server nvme reset

Required-githooks: true
