+++ src/control/cmd/daos_server/storage_scm.go
@@ -1,5 +1,5 @@
 //
-// (C) Copyright 2022 Intel Corporation.
+// (C) Copyright 2022-2023 Intel Corporation.
 //
 // SPDX-License-Identifier: BSD-2-Clause-Patent
 //
@@ -7,27 +7,16 @@
 package main
 
 import (
-	"context"
 	"strings"
 
 	"github.com/pkg/errors"
 
 	"github.com/daos-stack/daos/src/control/cmd/dmg/pretty"
 	"github.com/daos-stack/daos/src/control/common"
-	"github.com/daos-stack/daos/src/control/common/cmdutil"
-	"github.com/daos-stack/daos/src/control/lib/hardware"
-	"github.com/daos-stack/daos/src/control/lib/hardware/hwprov"
-	"github.com/daos-stack/daos/src/control/logging"
 	"github.com/daos-stack/daos/src/control/server"
 	"github.com/daos-stack/daos/src/control/server/config"
-	"github.com/daos-stack/daos/src/control/server/engine"
 	"github.com/daos-stack/daos/src/control/server/storage"
 )
-
-type scmSocketCmd struct {
-	affinitySource config.EngineAffinityFn
-	SocketID       *uint `long:"socket" description:"Perform PMem namespace operations on the socket identified by this ID (defaults to all sockets). PMem region operations will be performed across all sockets."`
-}
 
 const MsgStoragePrepareWarn = "Memory allocation goals for PMem will be changed and namespaces " +
 	"modified, this may be a destructive operation. Please ensure namespaces are unmounted " +
@@ -36,67 +25,16 @@ const MsgStoragePrepareWarn = "Memory allocation goals for PMem will be changed 
 
 type scmPrepareResetFn func(storage.ScmPrepareRequest) (*storage.ScmPrepareResponse, error)
 
-type scmCmd struct {
+type scmStorageCmd struct {
 	Prepare prepareSCMCmd `command:"prepare" description:"Prepare SCM devices so that they can be used with DAOS"`
 	Reset   resetSCMCmd   `command:"reset" description:"Reset SCM devices that have been used with DAOS"`
 	Scan    scanSCMCmd    `command:"scan" description:"Scan SCM devices"`
 }
 
 type prepareSCMCmd struct {
-	cmdutil.LogCmd `json:"-"`
-	helperLogCmd   `json:"-"`
-	optCfgCmd      `json:"-"`
-	scmSocketCmd   `json:"-"`
-
+	scmCmd
 	NrNamespacesPerSocket uint `short:"S" long:"scm-ns-per-socket" description:"Number of PMem namespaces to create per socket" default:"1"`
 	Force                 bool `short:"f" long:"force" description:"Perform SCM operations without waiting for confirmation"`
-}
-
-func setSockFromCfg(log logging.Logger, cfg *config.Server, affSrc config.EngineAffinityFn, req *storage.ScmPrepareRequest) error {
-	msgSkip := "so skip fetching sockid from config"
-
-	if cfg == nil {
-		log.Debug("nil input server config " + msgSkip)
-		return nil
-	}
-	if req.SocketID != nil {
-		log.Debug("sockid set in req " + msgSkip)
-		return nil
-	}
-
-	// Set engine NUMA affinities accurately in config.
-	if err := cfg.SetEngineAffinities(log, affSrc); err != nil {
-		return errors.Wrapf(err, "failed to set engine affinities %s: %s", msgSkip, err)
-	}
-
-	socksToPrep := make(map[uint]bool)
-
-	for _, ec := range cfg.Engines {
-		for _, scmCfg := range ec.Storage.Tiers.ScmConfigs() {
-			if scmCfg.Class == storage.ClassRam {
-				continue
-			}
-			socksToPrep[ec.Storage.NumaNodeIndex] = true
-		}
-	}
-
-	// Only set socket ID in request if one engine has been configured in config file with SCM
-	// storage specified, if engines exist assigned to different sockets, don't set the value.
-
-	switch len(socksToPrep) {
-	case 0:
-	case 1:
-		log.Debug("config contains scm-engines attached to a single socket, " +
-			"only process that socket")
-		for k := range socksToPrep {
-			req.SocketID = &k
-		}
-	default:
-		log.Debug("config contains scm-engines attached to different sockets, " +
-			"process all sockets")
-	}
-
-	return nil
 }
 
 func (cmd *prepareSCMCmd) preparePMem(prepareBackend scmPrepareResetFn) error {
@@ -115,19 +53,13 @@ func (cmd *prepareSCMCmd) preparePMem(prepareBackend scmPrepareResetFn) error {
 		SocketID:              cmd.SocketID,
 		NrNamespacesPerSocket: cmd.NrNamespacesPerSocket,
 	}
-
-	if err := setSockFromCfg(cmd.Logger, cmd.config, cmd.affinitySource, &req); err != nil {
-		return errors.Wrap(err, "setting sockid in prepare request")
-	}
-
-	cmd.Debugf("scm prepare request parameters: %+v", req)
+	cmd.Debugf("scm prepare request: %+v", req)
 
 	// Prepare PMem modules to be presented as pmem device files.
 	resp, err := prepareBackend(req)
 	if err != nil {
 		return err
 	}
-
 	cmd.Debugf("scm prepare response: %+v", resp)
 
 	if resp == nil {
@@ -176,43 +108,10 @@ func (cmd *prepareSCMCmd) preparePMem(prepareBackend scmPrepareResetFn) error {
 	return nil
 }
 
-func genFiAffFn(fis *hardware.FabricInterfaceSet) config.EngineAffinityFn {
-	return func(l logging.Logger, e *engine.Config) (uint, error) {
-		fi, err := fis.GetInterfaceOnNetDevice(e.Fabric.Interface, e.Fabric.Provider)
-		if err != nil {
-			return 0, err
-		}
-		return fi.NUMANode, nil
-	}
-}
-
-func getAffinitySource(log logging.Logger) (config.EngineAffinityFn, error) {
-	scanner := hwprov.DefaultFabricScanner(log)
-
-	fiSet, err := scanner.Scan(context.Background())
-	if err != nil {
-		return nil, errors.Wrap(err, "scan fabric")
-	}
-
-	return genFiAffFn(fiSet), nil
-}
-
-func (cmd *prepareSCMCmd) Execute(args []string) error {
-	if err := cmd.setHelperLogFile(); err != nil {
+func (cmd *prepareSCMCmd) Execute(_ []string) error {
+	if err := cmd.init(); err != nil {
 		return err
 	}
-
-	if cmd.IgnoreConfig {
-		cmd.config = nil
-	} else {
-		affSrc, err := getAffinitySource(cmd.Logger)
-		if err != nil {
-			cmd.Error(err.Error())
-		} else {
-			cmd.affinitySource = affSrc
-		}
-	}
-
 	scs := server.NewStorageControlService(cmd.Logger, config.DefaultServer().Engines)
 
 	cmd.Debugf("executing prepare scm command: %+v", cmd)
@@ -220,11 +119,7 @@ func (cmd *prepareSCMCmd) Execute(args []string) error {
 }
 
 type resetSCMCmd struct {
-	cmdutil.LogCmd `json:"-"`
-	helperLogCmd   `json:"-"`
-	optCfgCmd      `json:"-"`
-	scmSocketCmd   `json:"-"`
-
+	scmCmd
 	Force bool `short:"f" long:"force" description:"Perform PMem prepare operation without waiting for confirmation"`
 }
 
@@ -240,19 +135,13 @@ func (cmd *resetSCMCmd) resetPMem(resetBackend scmPrepareResetFn) error {
 		SocketID: cmd.SocketID,
 		Reset:    true,
 	}
-
-	if err := setSockFromCfg(cmd.Logger, cmd.config, cmd.affinitySource, &req); err != nil {
-		return errors.Wrap(err, "setting sockid in prepare (reset) request")
-	}
-
-	cmd.Debugf("scm prepare (reset) request parameters: %+v", req)
+	cmd.Debugf("scm prepare (reset) request: %+v", req)
 
 	// Reset PMem modules to default memory mode after removing any PMem namespaces.
 	resp, err := resetBackend(req)
 	if err != nil {
 		return err
 	}
-
 	cmd.Debugf("scm prepare (reset) response: %+v", resp)
 
 	state := resp.Socket.State
@@ -289,22 +178,10 @@ func (cmd *resetSCMCmd) resetPMem(resetBackend scmPrepareResetFn) error {
 	return nil
 }
 
-func (cmd *resetSCMCmd) Execute(args []string) error {
-	if err := cmd.setHelperLogFile(); err != nil {
+func (cmd *resetSCMCmd) Execute(_ []string) error {
+	if err := cmd.init(); err != nil {
 		return err
 	}
-
-	if cmd.IgnoreConfig {
-		cmd.config = nil
-	} else {
-		affSrc, err := getAffinitySource(cmd.Logger)
-		if err != nil {
-			cmd.Error(err.Error())
-		} else {
-			cmd.affinitySource = affSrc
-		}
-	}
-
 	scs := server.NewStorageControlService(cmd.Logger, config.DefaultServer().Engines)
 
 	cmd.Debugf("executing remove namespaces command: %+v", cmd)
@@ -312,37 +189,38 @@ func (cmd *resetSCMCmd) Execute(args []string) error {
 }
 
 type scanSCMCmd struct {
-	cmdutil.LogCmd `json:"-"`
-	helperLogCmd   `json:"-"`
-	optCfgCmd      `json:"-"`
+	scmCmd
 }
 
-func (cmd *scanSCMCmd) Execute(args []string) error {
-	var bld strings.Builder
-
-	if err := cmd.setHelperLogFile(); err != nil {
+func (cmd *scanSCMCmd) Execute(_ []string) error {
+	if err := cmd.init(); err != nil {
 		return err
 	}
-
 	svc := server.NewStorageControlService(cmd.Logger, config.DefaultServer().Engines)
 
 	cmd.Info("Scanning locally-attached PMem storage...\n")
 
-	pmemResp, err := svc.ScmScan(storage.ScmScanRequest{})
+	req := storage.ScmScanRequest{
+		SocketID: cmd.SocketID,
+	}
+	cmd.Debugf("scm scan request: %+v", req)
+
+	resp, err := svc.ScmScan(req)
 	if err != nil {
 		return err
 	}
+	cmd.Debugf("scm scan response: %+v", resp)
 
-	if len(pmemResp.Namespaces) > 0 {
-		if err := pretty.PrintScmNamespaces(pmemResp.Namespaces, &bld); err != nil {
+	var bld strings.Builder
+	if len(resp.Namespaces) > 0 {
+		if err := pretty.PrintScmNamespaces(resp.Namespaces, &bld); err != nil {
 			return err
 		}
 	} else {
-		if err := pretty.PrintScmModules(pmemResp.Modules, &bld); err != nil {
+		if err := pretty.PrintScmModules(resp.Modules, &bld); err != nil {
 			return err
 		}
 	}
-
 	cmd.Info(bld.String())
 
 	return nil
